Update to location of function, and to FAQ url
