+++ frontend/src/store/slice/Pipeline/PipelineHook.ts
@@ -1,14 +1,8 @@
-import {
-  useCallback,
-  useEffect,
-  useState,
-  createElement,
-  MouseEvent,
-} from "react"
+import { useCallback, useEffect, useState } from "react"
 import { useDispatch, useSelector } from "react-redux"
 import { useLocation, useNavigate, useParams } from "react-router-dom"
 
-import { OptionsObject, SnackbarKey, useSnackbar, VariantType } from "notistack"
+import { useSnackbar, VariantType } from "notistack"
 
 import { isRejected } from "@reduxjs/toolkit"
